+++ app_cv/test/v2/cv_model_test.dart
@@ -1,5 +1,10 @@
+/*
 import 'package:tekartik_app_cv/app_cv.dart';
 import 'package:tekartik_app_cv/src/builder.dart';
+
+ */
+import 'package:cv/src/builder.dart';
+import 'package:tekartik_app_cv/app_cv_v2.dart';
 import 'package:tekartik_common_utils/common_utils_import.dart';
 import 'package:test/test.dart';
 
@@ -180,6 +185,7 @@ void main() {
       } on UnsupportedError catch (e) {
         print(e);
       }
+      /*
       try {
         WithDuplicatedCvFields().fromModel({});
         fail('should fail');
@@ -191,7 +197,7 @@ void main() {
         fail('should fail');
       } on UnsupportedError catch (e) {
         print(e);
-      }
+      }*/
     });
     test('content child', () {
       expect(WithChildCvField().toModel(), {});