+++ app_cv_firestore/lib/app_cv_firestore_v1.dart
@@ -1,0 +1,24 @@
+/// Firestore content values helper
+library tekartik_app_cv_firestore_v1;
+
+export 'src/builder.dart'
+    show
+        CvFirestoreDocument,
+        CvFirestoreDocumentBase,
+        cvFirestoreAddBuilder,
+        CvFirestoreMapDocument,
+        CvFirestoreDocumentExt;
+export 'src/cv_collection_reference.dart' show CvCollectionReference;
+export 'src/cv_document_reference.dart' show CvDocumentReference;
+export 'src/cv_query_reference.dart' show CvQueryReference;
+export 'src/firestore_extension.dart'
+    show
+        CvFirestoreDocumentSnapshotExt,
+        CvFirestoreDocumentSnapshotsExt,
+        CvFirestoreQuerySnapshotExt,
+        CvFirestoreExt,
+        CvFirestoreCollectionReferenceExt,
+        CvFirestoreDocumentReferenceExt,
+        CvFirestoreQueryExt,
+        CvFirestoreWriteBatch,
+        CvFirestoreTransaction;