+++ app_cv_firestore/lib/app_cv_firestore_v2.dart
@@ -1,0 +1,26 @@
+/// Firestore content values helper
+library tekartik_app_cv_firestore_v2;
+
+export 'src/v2/builder.dart'
+    show
+        CvFirestoreDocument,
+        CvFirestoreDocumentBase,
+        cvFirestoreAddBuilder,
+        CvFirestoreMapDocument,
+        CvFirestoreDocumentExt;
+export 'src/v2/cv_collection_reference.dart' show CvCollectionReference;
+export 'src/v2/cv_document_reference.dart' show CvDocumentReference;
+export 'src/v2/cv_query_reference.dart' show CvQueryReference;
+export 'src/v2/firestore_extension.dart'
+    show
+        CvFirestoreDocumentSnapshotExt,
+        CvFirestoreDocumentSnapshotsExt,
+        CvFirestoreQuerySnapshotExt,
+        CvFirestoreExt,
+        CvFirestoreCollectionReferenceExt,
+        CvFirestoreDocumentReferenceExt,
+        CvFirestoreQueryExt,
+        CvFirestoreWriteBatch,
+        CvFirestoreTransaction;
+export 'src/v2/map_extension.dart'
+    show AppCvFirestoreFieldMapExt, WithServerTimestampMixin;