+++ app_cv_firestore/lib/src/v2/cv_collection_reference.dart
@@ -1,8 +1,9 @@
 import 'package:path/path.dart';
-import 'package:tekartik_app_cv_firestore/app_cv_firestore.dart';
-import 'package:tekartik_app_cv_firestore/src/cv_document_reference.dart';
-import 'package:tekartik_app_cv_firestore/src/cv_query_reference.dart';
+import 'package:tekartik_app_cv_firestore/app_cv_firestore_v2.dart';
 import 'package:tekartik_firebase_firestore/firestore.dart';
+
+import 'cv_document_reference.dart';
+import 'cv_query_reference.dart';
 
 /// Collection helper
 class CvCollectionReference<T extends CvFirestoreDocument> {