+++ app_csv/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.2
+
+- Fix nullValue in mapListToCsv
+
 ## 1.0.0
 
-- Initial version, created by Stagehand
+- Initial version