+++ app_cv/CHANGELOG.md
@@ -1,3 +1,7 @@
-## 1.0.0
+## 0.5.0
 
-- Initial version, created by Stagehand
+- Add `CvField.withName`
+
+## 0.1.0
+
+- Initial version